Extract ternary inner/outer logic as separate Dashboard/Project components
